NaiveBufferMgr cut off as module
