+++ src/buffer/manager/naive.rs
@@ -1,0 +1,129 @@
+use std::{
+    sync::{Arc, Mutex},
+    thread,
+    time::{Duration, SystemTime},
+};
+
+use anyhow::Result;
+
+use crate::{
+    buffer::buffer::Buffer,
+    file::{block_id::BlockId, manager::FileMgr},
+    log::manager::LogMgr,
+};
+
+use super::{BufferMgr, BufferMgrError, MAX_TIME};
+
+#[derive(Debug, Clone)]
+pub struct NaiveBufferMgr {
+    bufferpool: Vec<Arc<Mutex<Buffer>>>,
+    num_available: Arc<Mutex<usize>>,
+}
+
+impl NaiveBufferMgr {
+    pub fn new(fm: Arc<Mutex<FileMgr>>, lm: Arc<Mutex<LogMgr>>, numbuffs: usize) -> Self {
+        let bufferpool = (0..numbuffs)
+            .map(|_| Arc::new(Mutex::new(Buffer::new(Arc::clone(&fm), Arc::clone(&lm)))))
+            .collect();
+
+        Self {
+            bufferpool,
+            num_available: Arc::new(Mutex::new(numbuffs)),
+        }
+    }
+    // TODO: fix for thread safe
+    fn try_to_pin(&mut self, blk: &BlockId) -> Result<Arc<Mutex<Buffer>>> {
+        let mut buff = self.find_existing_buffer(blk);
+        if buff.is_none() {
+            buff = self.choose_unpinned_buffer();
+            if buff.is_none() {
+                return Err(From::from(BufferMgrError::BufferAbort));
+            }
+            let mut b = buff.as_ref().unwrap().lock().unwrap();
+            b.assign_to_block(blk.clone())?;
+        }
+
+        let mut b = buff.as_ref().unwrap().lock().unwrap();
+        if !b.is_pinned() {
+            *(self.num_available.lock().unwrap()) -= 1;
+        }
+        b.pin();
+
+        drop(b); // release lock
+        Ok(Arc::clone(&buff.unwrap()))
+    }
+    fn find_existing_buffer(&self, blk: &BlockId) -> Option<Arc<Mutex<Buffer>>> {
+        for i in 0..self.bufferpool.len() {
+            let buff = self.bufferpool[i].lock().unwrap();
+
+            if let Some(b) = buff.block() {
+                if *b == *blk {
+                    return Some(Arc::clone(&self.bufferpool[i]));
+                }
+            }
+        }
+
+        None
+    }
+    // The Naive Strategy
+    fn choose_unpinned_buffer(&mut self) -> Option<Arc<Mutex<Buffer>>> {
+        for i in 0..self.bufferpool.len() {
+            let buff = self.bufferpool[i].lock().unwrap();
+
+            if !buff.is_pinned() {
+                return Some(Arc::clone(&self.bufferpool[i]));
+            }
+        }
+
+        None
+    }
+}
+impl BufferMgr for NaiveBufferMgr {
+    // synchronized
+    fn available(&self) -> usize {
+        *(self.num_available.lock().unwrap())
+    }
+    // synchronized
+    fn flush_all(&mut self, txnum: i32) -> Result<()> {
+        for i in 0..self.bufferpool.len() {
+            let mut buff = self.bufferpool[i].lock().unwrap();
+            if buff.modifying_tx() == txnum {
+                buff.flush()?;
+            }
+        }
+
+        Ok(())
+    }
+    // synchronized
+    fn unpin(&mut self, buff: Arc<Mutex<Buffer>>) -> Result<()> {
+        let mut b = buff.lock().unwrap();
+
+        b.unpin();
+
+        if !b.is_pinned() {
+            *(self.num_available.lock().unwrap()) += 1;
+        }
+
+        Ok(())
+    }
+    // synchronized
+    fn pin(&mut self, blk: &BlockId) -> Result<Arc<Mutex<Buffer>>> {
+        let timestamp = SystemTime::now();
+
+        while !waiting_too_long(timestamp) {
+            if let Ok(buff) = self.try_to_pin(blk) {
+                return Ok(buff);
+            }
+            thread::sleep(Duration::new(1, 0))
+        }
+
+        return Err(From::from(BufferMgrError::BufferAbort));
+    }
+}
+
+fn waiting_too_long(starttime: SystemTime) -> bool {
+    let now = SystemTime::now();
+    let diff = now.duration_since(starttime).unwrap();
+
+    diff.as_millis() as i64 > MAX_TIME
+}
